Actually clean up anchor component
