--- mobile/components/Anchor/MessageAnchor.js
@@ -1,8 +1,0 @@
-// @flow
-import Anchor from './ThreadAnchor';
-import styled from 'styled-components';
-
-export const MessageAnchor = styled(Anchor)`
-  color: ${props => props.theme.text.default};
-  font-weight: 700;
-`;
